requirements.txt
absl-py==0.9.0
apt-clone==0.2.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
beautifulsoup4==4.6.0
blinker==1.4
Brlapi==0.6.6
cachetools==4.1.0
certifi==2018.1.18
chardet==3.0.4
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.1.2
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
feedparser==5.2.1
future==0.17.1
gast==0.2.2
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphsurgeon==0.4.4
grpcio==1.30.0
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
importlib-metadata==1.7.0
Jetson.GPIO==2.0.8
joblib==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.2
MarkupSafe==1.0
matplotlib==3.3.0
mock==3.0.5
numpy==1.16.1
oauth==1.0.1
oauthlib==3.1.0
opt-einsum==3.2.1
PAM==0.4.2
pandas==0.22.0
pbr==5.4.5
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.6
scikit-learn==0.23.1
scipy==1.4.1
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
sklearn==0.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==2.1.1
tensorflow==2.1.0+nv20.4
tensorflow-estimator==2.1.0
tensorrt==7.1.0.16
termcolor==1.1.0
testresources==2.0.1
threadpoolctl==2.1.0
ubuntu-drivers-common==0.0.0
uff==0.6.7
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
urwid==2.0.1
wadllib==1.3.2
webencodings==0.5
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
